Implement UX updates
